Refactor codebase (#108)
* Move templates and static dirs to base

* Update usage section [README]

* Not using prettier anymore

* Split template tags into separate files

* Use official bootstrap cdn

* Clean up syntax in consumers

* Fix UnboundLocalError in consumers

* Update url patterns for apps

* Move username update to separate form

* Remove unused email template

* Move django-extensions to dev group

* Clean asgi config

* Revert account view back to validate both forms

* Update templates and styles

* Add autofocus to username input in chat form

* Update footer copyright year

* Add delete chat confirmation modal

* Store pfp pics in media storage instead of the db

* Store chat images in media storage instead of the db

* Add missing prod env variables

* Update logo, notice, features, db schema, usage [README]

* Fix notice formatting [README]

* Add authorization check to delete_message view

* Update tests for profile model

* Update tests for users views

* Update tests for users models

* Update tests for chatterbox views

* Update tests for consumers

* Remove unused chat variables in create_chat view

* Adjust chat box width and height

+++ chatterbox/views.py
@@ -1,22 +1,18 @@
 from django.shortcuts import render, redirect, get_object_or_404
+from django.http import HttpResponseRedirect, HttpResponseForbidden, JsonResponse
+from django.views.decorators.http import require_POST
 from django.contrib.auth.decorators import login_required
-from django.http import HttpResponseRedirect
-from django.urls import reverse
-from django.contrib import messages
 from django.contrib.auth.models import User
+from django.contrib import messages
+from django.urls import reverse
 from .models import PrivateChat, Message
 from .forms import PrivateChatForm
 
-
 @login_required
 def chat_list(request):
-    # Get all private chats involving the current user
-    chats = PrivateChat.objects.filter(user1=request.user) | PrivateChat.objects.filter(
-        user2=request.user
-    )
-
-    # Fetch the last message and time diff for each chat
+    chats = PrivateChat.objects.filter(user1=request.user) | PrivateChat.objects.filter(user2=request.user)
     chats_with_last_messages = []
+
     for chat in chats:
         last_message = chat.messages.order_by("-timestamp").first()
         time_diff = last_message.get_time_diff() if last_message else "No messages yet"
@@ -25,7 +21,6 @@ def chat_list(request):
 
     context = {"chats_with_last_messages": chats_with_last_messages}
     return render(request, "chatterbox/chat_list.html", context)
-
 
 @login_required
 def create_chat(request):
@@ -36,39 +31,25 @@ def create_chat(request):
             other_user = User.objects.filter(username=username).first()
 
             if not other_user:
-                messages.success(
-                    request,
-                    "The username you entered does not exist. Please try again.",
-                )
+                messages.error(request, "Username doesn't exist.")
                 return redirect("chatterbox:create_chat")
 
             if other_user == request.user:
-                messages.success(
-                    request, "You cannot start a chat with yourself. Please try again."
-                )
+                messages.success(request, "You can't start a chat with yourself.")
                 return redirect("chatterbox:create_chat")
 
-            # Fetch or create the chat
-            chat, _ = (
+            # fetch/create chat
+            if request.user.id < other_user.id:
                 PrivateChat.objects.get_or_create(user1=request.user, user2=other_user)
-                if request.user.id < other_user.id
-                else PrivateChat.objects.get_or_create(
-                    user1=other_user, user2=request.user
-                )
-            )
+            else:
+                PrivateChat.objects.get_or_create(user1=other_user, user2=request.user)
 
-            return HttpResponseRedirect(
-                reverse(
-                    "chatterbox:chat", args=[request.user.username, other_user.username]
-                )
-            )
+            return HttpResponseRedirect(reverse("chatterbox:chat", args=[request.user.username, other_user.username]))
 
     else:
         form = PrivateChatForm()
 
-    context = {"form": form}
-    return render(request, "chatterbox/create_chat.html", context)
-
+    return render(request, "chatterbox/create_chat.html", {"form": form})
 
 @login_required
 def chat(request, username, other_username):
@@ -79,67 +60,72 @@ def chat(request, username, other_username):
         return redirect("chatterbox:chat_list")
 
     if request.user == user1:
-        # Fetch or create the chat
-        chat, _ = (
-            PrivateChat.objects.get_or_create(user1=user1, user2=user2)
-            if user1.id < user2.id
-            else PrivateChat.objects.get_or_create(user1=user2, user2=user1)
-        )
+        # fetch/create chat
+        if user1.id < user2.id:
+            chat, _ = PrivateChat.objects.get_or_create(user1=user1, user2=user2)
+        else:
+            chat, _ = PrivateChat.objects.get_or_create(user1=user2, user2=user1)
     else:
         return redirect("chatterbox:chat_list")
 
     messages = Message.objects.filter(chat=chat).order_by("timestamp")
 
-    # Zip with previous message
+    # zip with previous message
     messages_with_prev = []
     prev = None
     for message in messages:
         messages_with_prev.append((message, prev))
         prev = message
 
-    context = {
-        "messages_with_prev": messages_with_prev,
-        "current_user": user1,
-        "other_user": user2,
-    }
+    context = {"messages_with_prev": messages_with_prev, "current_user": user1, "other_user": user2}
     return render(request, "chatterbox/chat.html", context)
 
+@login_required
+@require_POST
+def upload_image(request, username, other_username):
+    image = request.FILES.get("image")
+    if not image:
+        return JsonResponse({"error": "No image provided."}, status=400)
+
+    message = Message(user=request.user, chat_id=0)
+    message.image.save(image.name, image, save=False)
+    return JsonResponse({"image_name": message.image.name, "image_url": message.image.url})
 
 @login_required
 def delete_chat(request, username, other_username):
     try:
         user1 = User.objects.get(username=username)
         user2 = User.objects.get(username=other_username)
-        # Determine the correct order for user1 and user2 in the chat
-        chat = (
-            PrivateChat.objects.get(user1=user1, user2=user2)
-            if user1.id < user2.id
-            else PrivateChat.objects.get(user1=user2, user2=user1)
-        )
+
+        if request.user != user1 and request.user != user2:
+            messages.error(request, "Not authorized to delete this chat.")
+            return redirect("chatterbox:chat_list")
+
+        if user1.id < user2.id:
+            chat = PrivateChat.objects.get(user1=user1, user2=user2)
+        else:
+            chat = PrivateChat.objects.get(user1=user2, user2=user1)
         chat.delete()
-        messages.success(request, "Chat has been successfully deleted!")
+        messages.success(request, "Chat has been successfully deleted.")
+
     except User.DoesNotExist:
         messages.error(request, "One or both users do not exist.")
     except PrivateChat.DoesNotExist:
         messages.error(request, "Chat does not exist.")
-    except Exception as err:
+    except Exception:
         messages.error(request, "An unexpected error occurred.")
-        print(f"Unexpected error: {err}")
 
     return redirect("chatterbox:chat_list")
 
-
 @login_required
+@require_POST
 def delete_message(request, message_id):
     message = get_object_or_404(Message, id=message_id)
-    message.delete()
+    if message.user != request.user:
+        return HttpResponseForbidden("Not allowed to delete this message.")
 
     chat = message.chat
-    current_user = request.user
     other_user = chat.user2 if chat.user1 == request.user else chat.user1
+    message.delete()
 
-    return redirect(
-        "chatterbox:chat",
-        username=current_user.username,
-        other_username=other_user.username,
-    )
+    return redirect("chatterbox:chat", username=request.user.username, other_username=other_user.username)
